// packages/js/data/src/crud/action-types.ts
export enum TYPES {
	CREATE_ITEM_ERROR = 'CREATE_ITEM_ERROR',
	CREATE_ITEM_SUCCESS = 'CREATE_ITEM_SUCCESS',
	GET_ITEM_ERROR = 'GET_ITEM_ERROR',
	GET_ITEM_SUCCESS = 'GET_ITEM_SUCCESS',
	GET_ITEMS_ERROR = 'GET_ITEMS_ERROR',
	GET_ITEMS_SUCCESS = 'GET_ITEMS_SUCCESS',
	UPDATE_ITEM_ERROR = 'UPDATE_ITEM_ERROR',
	UPDATE_ITEM_SUCCESS = 'UPDATE_ITEM_SUCCESS',
}

export default TYPES;